move highfive event into own func export
